Various changes
Sounds are now stored in an array to allow for multiple instances of the same sound to both play at once. Levels have been moved to levels.txt and replaced with a test level. Created a file for enemies and created an enemy called strotter, added collisions to it and fine tuned it.